Example Mifos plugin using Groovy
The Java code is just a wrapper which executes a Groovy script. The
point is that Groovy can be used to do the real parsing work.

One limitation that comes to mind: I'm not sure if arbitrary Java
libraries may be used by the Groovy script. It's probably possible to
run other Groovy, but maybe not with dependencies like you'd get with a
pure Java plugin.

I also thought through another approach: running Groovy directly from
Mifos. This should work too. On the plus side, this makes a Java wrapper
unnecessary. I didn't implement this way because I really wanted to be
able to quickly change and run the script repeatedly.

Yet another option is simply compiling Groovy code: Mifos would run it
just like a pure Java plugin. Again, a quick edit-run-fix cycle is lost.

Perhaps OSGi is the answer here: that should allow plugins to be loaded
and removed at runtime.